requirements.txt
torch
Pillow
diffusers
accelerate
google-generativeai
tqdm
cache_dit
optimum-quanto